Move dynamic layout shape test into cache tests
Keep the large-shape packing coverage with the existing TensorAdaptor cache signature tests instead of adding a standalone test module.
