+++ tests/unit/test_tensor_cache_signature.py
@@ -11,11 +11,26 @@ kernel's compiled IR depends on concrete shape values, the affected dims must
 be marked with ``mark_static`` so they participate in the cache key.
 """
 
+import ctypes
+import struct
+
 import pytest
 import torch
 
 import flydsl.compiler as flyc
 from flydsl.compiler.jit_argument import TensorAdaptor
+
+
+class _FakeLargeTensor:
+    dtype = torch.float16
+    shape = (2**31 + 1,)
+
+    def data_ptr(self):
+        return 0x1234
+
+    def stride(self, dim=None):
+        strides = (1,)
+        return strides if dim is None else strides[dim]
 
 # -----------------------------------------------------------------------------
 # Default cache behavior: shape NOT in key, kernels reused across shapes
@@ -58,6 +73,24 @@ def test_raw_cache_signature_differs_by_rank():
     a = torch.empty((10,), dtype=torch.float32)
     b = torch.empty((2, 5), dtype=torch.float32)
     assert TensorAdaptor.raw_cache_signature(a) != TensorAdaptor.raw_cache_signature(b)
+
+
+def test_dynamic_layout_packs_large_shape_dim_as_i64():
+    tensor = _FakeLargeTensor()
+    spec = TensorAdaptor._reusable_slot_spec(tensor)
+
+    data_slot, layout_slot = spec
+    data_ctype, data_extract = data_slot
+    layout_ctype, layout_extract = layout_slot
+
+    assert data_ctype is ctypes.c_void_p
+    assert data_extract(tensor) == tensor.data_ptr()
+    assert ctypes.sizeof(layout_ctype) == 8
+
+    storage = layout_ctype()
+    layout_extract(tensor, storage)
+
+    assert struct.unpack_from("<q", memoryview(storage).cast("b"), 0)[0] == tensor.shape[0]
 
 
 # -----------------------------------------------------------------------------
